pc_app/requirements.txt
PyQt6
opencv-python
mss
numpy
requests
qrcode
pillow
google-cloud-firestore
google-auth
google-auth-oauthlib
